requirements.txt
RPi.GPIO
# smbus  This is provided by a raspbian package for now.
docopt
